fix: address riverside review feedback on session view PR
- Move session_render.rs inline tests to session_render_tests.rs
  following the project's test file placement convention (CLAUDE.md)
- Update README to document new 'session view' command and --watch flag
- Apply ANSI rendering in 'midtown coworker view' so rich-text output
  (tool calls as code fences) is rendered instead of raw markdown
- Delete dead session_view() client method (replaced by session_view_raw)
- Update get_output() doc comment to reflect rich-text format
- Clarify --watch help text: headless sessions only
